Blocked out Bottom Section on Public Campaign and Moved it into its own folder
Keep in mind that although it says that i deleted files, it was just moving them. I could have done git mv but I was too lazy.
