+++ src/App.js
@@ -8,7 +8,7 @@ import Register from "./components/homePage/3.register";
 import Overview from './components/Pages/Accounts/Campaign/Overview';
 import CreateProject from './components/Pages/Create-Project';
 import Discover from './components/Pages/Discover';
-import PublicCampaign from './components/Pages/PublicCampaign';
+import PublicCampaign from './components/Pages/PublicCampaign/PublicCampaign';
 import Search from './components/Pages/Search';
 
 const App = () => {
