+++ src/components/Pages/PublicCampaign/PublicCampaign.css
@@ -4,7 +4,7 @@
 	width: 100%;
 }
 
-.title {
+.campaign-title {
 	font-weight: bold;
 }
 
@@ -68,4 +68,27 @@ img {
 	width:100%;
 	height: 4rem;
 	font-weight: bold;
+}
+
+.bottom-section {
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+}
+
+.reward-title {
+	text-align: left;
+}
+
+.content-container {
+	width: 750px;
+}
+
+.rewards-container {
+	width: 450px;
+}
+
+.rewards-list{
+	display: flex;
+	flex-direction: column;
 }
